put formData struct in own file
